requirements.txt
# Core web framework
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20

# Authentication
msal==1.31.1
authlib==1.4.0
itsdangerous==2.2.0

# HTTP clients
requests==2.32.3
httpx==0.28.1

# AI/ML
openai==1.58.1

# Data processing
pandas==2.2.3
openpyxl==3.1.5
pdfplumber==0.11.4
python-docx==1.1.0
xlrd==2.0.1

# Utilities
python-dotenv==1.0.1
jinja2==3.1.5
apscheduler==3.10.4
aiofiles==24.1.0

# OCR support for scanned/image-based PDFs
pytesseract==0.3.13
pdf2image==1.17.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
greenlet==3.1.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
